fare-rule-normalization-and-automation-main/streamlit_version/src/pyproject.toml
[tool.poetry]
name = "amadeus_project"
version = "0.1.0"
description = ""
authors = ["tmelwig <thomas.melwig@student-cs.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^0.28.1"
langchain = "^0.0.331"
streamlit = "^1.28.1"
PyPDF2 = "^3.0.1"
python-utils = "^3.8.1"
pandas = "^2.1.2"
faiss-cpu = "^1.7.4"
tiktoken = "^0.5.1"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
